chore: break import cycles by moving shared utilities out of archives
Move eventKey.mjs to shared/utils/game/ and useScrollEvent.mjs to
shared/hooks/. Breaks bidirectional cycles between archives↔dashboard
and archives↔timeline. Both utilities are used by 3+ features.
