+++ src/features/archives/ArchivesClient.jsx
@@ -10,10 +10,10 @@ import EventLog from '@/features/timeline/EventLog';
 import ArchiveMap from '@/features/archives/ArchiveMap';
 import SeasonSelector from '@/features/archives/SeasonSelector';
 import RefreshSeasonButton from '@/features/archives/RefreshSeasonButton';
-import { eventKey } from '@/features/archives/eventKey.mjs';
+import { eventKey } from '@/shared/utils/game/eventKey.mjs';
 import { getWarOutcome } from '@/features/archives/getWarOutcome.mjs';
 import { useCyberstanEffects } from '@/features/archives/useCyberstanEffects.mjs';
-import { useScrollEvent } from '@/features/archives/useScrollEvent.mjs';
+import { useScrollEvent } from '@/shared/hooks/useScrollEvent.mjs';
 import { useHeaderGlassFilter } from '@/shared/hooks/useHeaderGlassFilter.mjs';
 import { usePersistedState } from '@/shared/hooks/usePersistedState.mjs';
 import { FACTION_KEY } from '@/shared/preferences/faction.mjs';
