+++ src/features/dashboard/HomeClient.jsx
@@ -7,8 +7,8 @@ import Galaxy from '@/features/galaxy/Galaxy';
 import DashboardClient from '@/features/dashboard/DashboardClient';
 import EventLog from '@/features/timeline/EventLog';
 import { useLiveDataContext } from '@/shared/providers/LiveDataContext.mjs';
-import { useScrollEvent } from '@/features/archives/useScrollEvent.mjs';
-import { eventKey } from '@/features/archives/eventKey.mjs';
+import { useScrollEvent } from '@/shared/hooks/useScrollEvent.mjs';
+import { eventKey } from '@/shared/utils/game/eventKey.mjs';
 import { computeMapStateAtEvent } from '@/shared/utils/game/computeMapStateAtEvent.mjs';
 import { computePulseDelays } from '@/shared/utils/game/pulseDelays.mjs';
 import { useHeaderGlassFilter } from '@/shared/hooks/useHeaderGlassFilter.mjs';
